+++ main.go
@@ -8,7 +8,7 @@ import (
 )
 
 func main() {
-	err := command.Build("0.9.1").Execute()
+	err := command.Build("0.10.0").Execute()
 	if err != nil {
 		fmt.Println(err)
 		os.Exit(1)